requirements.txt
tensorflow_addons
sklearn
numpy
pandas
argparse
PyWavelets